feat(tests): new test file for common service options
