+++ tests/mysql_test.go
@@ -12,8 +12,8 @@ import (
 	mysqluserconfig "github.com/aiven/aiven-operator/api/v1alpha1/userconfig/service/mysql"
 )
 
-func getMySQLYaml(project, name, cloudName string, includeTechnicalEmails bool) string {
-	baseYaml := `
+func getMySQLYaml(project, name, cloudName string) string {
+	return fmt.Sprintf(`
 apiVersion: aiven.io/v1alpha1
 kind: MySQL
 metadata:
@@ -39,16 +39,8 @@ spec:
       - network: 0.0.0.0/32
         description: bar
       - network: 10.20.0.0/16
-`
 
-	if includeTechnicalEmails {
-		baseYaml += `
-  technicalEmails:
-    - email: "test@example.com"
-`
-	}
-
-	return fmt.Sprintf(baseYaml, project, name, cloudName)
+`, project, name, cloudName)
 }
 
 func TestMySQL(t *testing.T) {
@@ -58,7 +50,7 @@ func TestMySQL(t *testing.T) {
 	// GIVEN
 	ctx := context.Background()
 	name := randName("mysql")
-	yml := getMySQLYaml(testProject, name, testPrimaryCloudName, false)
+	yml := getMySQLYaml(testProject, name, testPrimaryCloudName)
 	s := NewSession(k8sClient, avnClient, testProject)
 
 	// Cleans test afterwards
@@ -119,49 +111,4 @@ func TestMySQL(t *testing.T) {
 	assert.NotEmpty(t, secret.Data["MYSQL_SSL_MODE"])
 	assert.NotEmpty(t, secret.Data["MYSQL_URI"])
 	assert.NotEmpty(t, secret.Data["MYSQL_REPLICA_URI"]) // business-4 has replica
-}
-
-func TestMySQLTechnicalEmails(t *testing.T) {
-	t.Parallel()
-	defer recoverPanic(t)
-
-	// GIVEN
-	ctx := context.Background()
-	name := randName("mysql")
-	yml := getMySQLYaml(testProject, name, testPrimaryCloudName, true)
-	s := NewSession(k8sClient, avnClient, testProject)
-
-	// Cleans test afterwards
-	defer s.Destroy()
-
-	// WHEN
-	// Applies given manifest
-	require.NoError(t, s.Apply(yml))
-
-	// Waits kube objects
-	ms := new(v1alpha1.MySQL)
-	require.NoError(t, s.GetRunning(ms, name))
-
-	// THEN
-	// Technical emails are set
-	msAvn, err := avnClient.Services.Get(ctx, testProject, name)
-	require.NoError(t, err)
-	assert.Len(t, ms.Spec.TechnicalEmails, 1)
-	assert.Equal(t, "test@example.com", msAvn.TechnicalEmails[0].Email)
-
-	// WHEN
-	// Technical emails are removed from manifest
-	updatedYml := getMySQLYaml(testProject, name, testPrimaryCloudName, false)
-
-	// Applies updated manifest
-	require.NoError(t, s.Apply(updatedYml))
-
-	// Waits kube objects
-	require.NoError(t, s.GetRunning(ms, name))
-
-	// THEN
-	// Technical emails are removed from service
-	msAvnUpdated, err := avnClient.Services.Get(ctx, testProject, name)
-	require.NoError(t, err)
-	assert.Empty(t, msAvnUpdated.TechnicalEmails)
 }
